+++ mypy.ini
@@ -1,2 +1,5 @@
 [mypy]
 exclude = tests/
+
+[mypy-pysrc.sync]
+implicit_reexport = True
