+++ pysrc/environment.py
@@ -4,4 +4,4 @@ Collection of runtime constants
 from pathlib import Path
 
 PROJECT_ROOT = Path(__file__).parent.parent
-OUT_DIR = PROJECT_ROOT / "data"
+OUT_DIR = PROJECT_ROOT / Path("data")
