+++ pysrc/requirements.txt
@@ -1,5 +1,5 @@
 # Prod
-dune-client==0.2.0
+dune-client==0.2.2b0
 psycopg2-binary==2.9.3
 python-dotenv>=0.20.0
 requests>=2.28.1
