+++ pysrc/sync/__init__.py
@@ -1,0 +1,2 @@
+"""Re-exported sync methods."""
+from .app_data import sync_app_data
